Make `test_reprostim_timesync-stimuli.sh` script to work in both default modes and on virtual screen with `xvfb` argument, #147.

+++ tools/ci/test_reprostim_timesync-stimuli.sh
@@ -1,5 +1,31 @@
 #!/bin/bash
 
+####################################################
+# help begin
+show_help() {
+  cat << EOF
+Usage: $0 [MODE]
+
+Arguments:
+  MODE           Mode to run the script in. Use 'xvfb' to run it on virtual
+                 screen or omit this argument to run on the current DISPLAY.
+
+Options:
+  -h, --help     Show this help message and exit
+
+Description:
+  Test CI/CD ReproStim timesync-stimuli script.
+EOF
+}
+
+# Check for help argument
+if [[ "$1" == "-h" || "$1" == "--help" ]]; then
+  show_help
+  exit 0
+fi
+# help end
+####################################################
+
 thisdir=$(dirname "$0")
 tmp_dir="${TMPDIR:-/tmp}"
 MODE=${1:-default}
@@ -7,46 +33,62 @@ MODE=${1:-default}
 echo "Test CI/CD ReproStim timesync-stimuli.."
 
 echo "tmp_dir=${tmp_dir}"
+cd ${thisdir}
 
 if [[ "$MODE" == "xvfb" ]]; then
   echo "Running in CI/CD mode on virtual screen"
   # store the tmp_dir in GITHUB_ENV for later use
   echo "tmp_dir=$tmp_dir" >> "$GITHUB_ENV"
+
+  export FRAME_WIDTH=1920
+  export FRAME_HEIGHT=108055
+  export FRAME_RATE=60
+  export FRAME_BPP=24
+  export DISPLAY_PATH="$tmp_dir/reprostim_last_display.txt"
+  export XVFB_OPTS="-screen 0 ${FRAME_WIDTH}x${FRAME_HEIGHT}x${FRAME_BPP} -ac +extension GLX +render -noreset"
+  export DISPLAY_START=25
+  export REPROSTIM_CMD="./run_reprostim_container.sh timesync-stimuli -m event --mute -d \$(cat $tmp_dir/reprostim_last_display.txt)"
+
+  echo "Run Xvfb in background with REPROSTIM_CMD"
+  xvfb-run -a -n $DISPLAY_START -s "$XVFB_OPTS" bash -c "echo \$DISPLAY > ${DISPLAY_PATH}; $REPROSTIM_CMD"&
+  XVFB_RUN_PID=$!
+
+  echo "Started xvfb-run with PID $XVFB_RUN_PID"
+  echo "Wait for Xvfb to start"
+  sleep 4
+
+  export DISPLAY_ID=$(cat ${DISPLAY_PATH})
+  echo "Xvfb started on display: $DISPLAY_ID"
 else
   echo "Running in default mode on current DISPLAY=$DISPLAY"
+
+  read resolution refresh_rate < <(xrandr | grep '*' | awk '{print $1, $2}')
+
+  export DISPLAY_ID="${DISPLAY#:}"
+  export FRAME_WIDTH=${resolution%x*}
+  export FRAME_HEIGHT=${resolution#*x}
+  export FRAME_RATE="${refresh_rate%%[^0-9.]*}"
+  export FRAME_BPP=24
+  export REPROSTIM_CMD="./run_reprostim_container.sh timesync-stimuli -m event --mute -d \${DISPLAY_ID}"
+
 fi
 
+echo "Display[$DISPLAY_ID]: ${FRAME_WIDTH}x${FRAME_HEIGHT}, ${FRAME_RATE}Hz"
+echo "ReproStim command to run: $REPROSTIM_CMD"
 
-export FRAME_WIDTH=1920
-export FRAME_HEIGHT=1080
-export FRAME_RATE=60
-export FRAME_BPP=24
-export DISPLAY_PATH="$tmp_dir/reprostim_last_display.txt"
-export XVFB_OPTS="-screen 0 ${FRAME_WIDTH}x${FRAME_HEIGHT}x${FRAME_BPP} -ac +extension GLX +render -noreset"
-export DISPLAY_START=25
-export REPROSTIM_CMD="./run_reprostim_container.sh timesync-stimuli -m event --mute -d \$(cat $tmp_dir/reprostim_last_display.txt)"
-
-cd ${thisdir}
-
-echo "Run Xvfb in background with REPROSTIM_CMD"
-xvfb-run -a -n $DISPLAY_START -s "$XVFB_OPTS" bash -c "echo \$DISPLAY > ${DISPLAY_PATH}; $REPROSTIM_CMD"&
-XVFB_RUN_PID=$!
-
-echo "Started xvfb-run with PID $XVFB_RUN_PID"
-echo "Wait for Xvfb to start"
-sleep 4
-
-export DISPLAY=$(cat ${DISPLAY_PATH})
-echo "Xvfb started on display: $DISPLAY"
 echo "Send test pulse events"
-./test_reprostim_events.sh 2 5 5 1.5 20 "${DISPLAY}" &
+./test_reprostim_events.sh 2 5 5 1.5 20 "${DISPLAY_ID}" &
 
 echo "Record video for 45 seconds"
-ffmpeg -video_size "${FRAME_WIDTH}x${FRAME_HEIGHT}" -framerate "${FRAME_RATE}" -f x11grab -i "$DISPLAY" -t 45 -c:v libx264 -pix_fmt yuv420p "$tmp_dir/reprostim_screenshot_$(date +%Y-%m-%d_%H-%M-%S).mp4"
+echo "ffmpeg -video_size \"${FRAME_WIDTH}x${FRAME_HEIGHT}\" -framerate \"${FRAME_RATE}\" -f x11grab -i \"$DISPLAY_ID\" -t 45 -c:v libx264 -pix_fmt yuv420p \"$tmp_dir/reprostim_screenshot_$(date +%Y-%m-%d_%H-%M-%S).mp4\""
+ffmpeg -video_size "${FRAME_WIDTH}x${FRAME_HEIGHT}" -framerate "${FRAME_RATE}" -f x11grab -i "$DISPLAY_ID" -t 45 -c:v libx264 -pix_fmt yuv420p "$tmp_dir/reprostim_screenshot_$(date +%Y-%m-%d_%H-%M-%S).mp4"
 sleep 45
 ls -l $tmp_dir/reprostim_*
 
-echo "Kill Xvfb at the end if any"
-sleep 1
-kill $XVFB_RUN_PID 2>/dev/null || true
-wait $XVFB_RUN_PID 2>/dev/null || true
+# terminate xvfb process if it was started
+if [[ "$MODE" == "xvfb" ]]; then
+  echo "Kill Xvfb at the end if any"
+  sleep 1
+  kill $XVFB_RUN_PID 2>/dev/null || true
+  wait $XVFB_RUN_PID 2>/dev/null || true
+fi
